wait for empty JS call stack before calling view command
